+++ .storybook/config.js
@@ -2,6 +2,7 @@ import { configure } from '@kadira/storybook';
 import '../styles/ColorPicker.scss';
 import '../styles/Switch.scss';
 import '../styles/TagsInput.scss';
+import '../styles/LoadingSpinner.scss';
 
 function loadStories() {
   require('../src/stories/Switch'),
@@ -9,6 +10,7 @@ function loadStories() {
   require('../src/stories/YoutubeVideo'),
   require('../src/stories/VimeoVideo'),
   require('../src/stories/ColorPicker');
+  require('../src/stories/LoadingSpinner');
 }
 
 configure(loadStories, module);