+++ src/stories/YoutubeVideo.js
@@ -5,5 +5,5 @@ import YoutubeVideo from '../YoutubeVideo';
 
 storiesOf('YoutubeVideo', module)
   .add('default', () => (
-    <YoutubeVideo src="https://www.youtube.com/embed/NK_e5_fKtIo" />
+    <YoutubeVideo src="https://www.youtube.com/watch?v=H5ud1oB4ei4" />
   ));