+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * Switched from grunt-react to grunt-babel.
 * Updated React dependency to v15.0.1.
 * Switch is now a controlled component. Previously it was uncontrolled, maintaining it's own state.
+* Added LoadingSpinner component